Fixing requested changes and restructuring
