+++ src/functions/isAutheticated.ts
@@ -1,10 +1,9 @@
 import admin from 'firebase-admin';
-import { Request, Response } from 'express';
+import { Request, Response, NextFunction } from 'express';
 import { responses } from '../constants/responses.js';
 import { statusCodes } from '../constants/codes.js';
 
-// eslint-disable-next-line @typescript-eslint/ban-types
-export const isAuthenticated = async(req: Request, res: Response, next: Function): Promise<Response> => {
+export const isAuthenticated = async(req: Request, res: Response, next: NextFunction): Promise<Response> => {
 	const { authorization } = req.headers;
 
 	if (!authorization || !authorization.startsWith('Bearer')) {
@@ -29,7 +28,7 @@ export const isAuthenticated = async(req: Request, res: Response, next: Function
 			email: decodedToken.email,
 		};
 
-		return next();
+		next();
 	} catch (error) {
 		return res.status(statusCodes.unauthorized).send({ message: error.message});
 	}
