+++ src/functions/isAuthorized.ts
@@ -1,14 +1,13 @@
-import { Request, Response } from 'express';
+import { Request, Response, NextFunction } from 'express';
 import { statusCodes } from '../constants/codes.js';
 import { responses } from '../constants/responses.js';
 
 export const isAuthorized = (opts: { hasRole: Array<'admin' | 'user'>, allowSameUser?: boolean }) => 
-	// eslint-disable-next-line @typescript-eslint/ban-types
-	 (req: Request, res: Response, next: Function) => {
+	 (req: Request, res: Response, next: NextFunction) => {
 		const { role, uid } = res.locals;
 		const { id } = req.params;
 
-		if (opts.allowSameUser && id && uid === id)
+		if (opts.allowSameUser && uid === id)
 			return next();
 
 		if (!role)
