+++ src/types/errorType.ts
@@ -1,0 +1,6 @@
+interface ErrorType {
+  code: string;
+  message: string;
+}
+
+export default ErrorType;
